+++ client/src/components/dashboard/StudentManagementSection.jsx
@@ -1,271 +1,18 @@
 // File: client/src/components/dashboard/StudentManagementSection.jsx
-// PHASE 1 REFACTORED: Training Modal Integration + Utility Functions Extracted
-// Lines reduced from 535 → 380 (-155 lines, -29% reduction)
-// PRESERVED: All existing functionality - zero breaking changes
-import React, { useState } from "react";
-import { ensureArray, normalizeStudentData, createSafeHandler } from "../../utils/studentTableUtils";
-import { getStatusConfig, tabsConfig } from "../../utils/studentStatusConfig";
-import StudentStatusDisplay from "../student/StudentStatusDisplay";
+// PHASE 2: Component extraction complete - 380 → 180 lines (-53% total)
+import React from "react";
+import { ensureArray } from "../../utils/studentTableUtils";
+import { tabsConfig } from "../../utils/studentStatusConfig";
+import StudentCard from "../student/StudentCard";
+import StudentTableRow from "../student/StudentTableRow";
+import SearchControls from "../student/SearchControls";
+import TabNavigation from "../student/TabNavigation";
 
-// Lines 10-85: Mobile Student Card Component - Refactored with extracted utilities
-const StudentCard = ({ 
-  student, 
-  onProcessPayment, 
-  onViewStudent, 
-  onEditStudent, 
-  onSendReminder, 
-  onLogTraining,
-  canSendReminder, 
-  smsLoading,
-  students = [],
-  getStudentStatus,
-  getDaysRemaining
-}) => {
-  if (!student || typeof student !== 'object') {
-    return null;
-  }
-
-  // Lines 25-30: Use utility function for data normalization
-  const studentData = normalizeStudentData(student);
-  const { name, email, phone, monthlyRate } = studentData;
-
-  // Lines 32-37: Status calculations
-  const status = getStudentStatus ? getStudentStatus(student) : 'inactive';
-  const daysText = getDaysRemaining ? getDaysRemaining(student) : 'Unknown';
-  const canSendSMS = canSendReminder ? canSendReminder(student) : false;
-
-  // Lines 39-47: Safe event handlers using utility
-  const handleEditClick = createSafeHandler(
-    () => onEditStudent(student), 
-    'Mobile Edit - onEditStudent'
-  );
-
-  const handleTrainingClick = createSafeHandler(
-    () => onLogTraining(student), 
-    'Mobile Training - onLogTraining'
-  );
-
-  return (
-    <div className="bg-gray-750 rounded-xl p-5 border border-gray-600 shadow-lg">
-      <div className="flex items-start justify-between mb-4">
-        <div className="flex items-center space-x-3 flex-1 min-w-0">
-          <div className="h-12 w-12 rounded-full bg-gradient-to-r from-blue-400 to-purple-500 flex items-center justify-center text-white font-bold text-lg flex-shrink-0">
-            {name.charAt(0).toUpperCase()}
-          </div>
-          <div className="flex-1 min-w-0">
-            <h3 className="text-white font-semibold text-base truncate">{name}</h3>
-            <p className="text-gray-400 text-sm truncate">{email}</p>
-            <p className="text-gray-400 text-sm">{phone}</p>
-          </div>
-        </div>
-        
-        {/* Lines 65-70: Use new status display component */}
-        <div className="flex-shrink-0 text-right space-y-1">
-          <StudentStatusDisplay 
-            status={status} 
-            student={student} 
-            showLegacy={true}
-          />
-        </div>
-      </div>
-
-      <div className="grid grid-cols-2 gap-4 mb-4">
-        <div>
-          <p className="text-xs text-gray-400 uppercase tracking-wider mb-1">Rate</p>
-          <p className="text-white font-semibold">₱{monthlyRate.toLocaleString()}</p>
-          <p className="text-gray-400 text-xs">per month</p>
-        </div>
-        <div>
-          <p className="text-xs text-gray-400 uppercase tracking-wider mb-1">Due Date</p>
-          <p className="text-sm font-medium text-gray-300">
-            {daysText}
-          </p>
-        </div>
-      </div>
-
-      <div className="space-y-3">
-        <div className="grid grid-cols-4 gap-2">
-          <button
-            onClick={() => onViewStudent && onViewStudent(student.id)}
-            className="flex flex-col items-center justify-center py-3 px-1 bg-blue-600 hover:bg-blue-700 active:bg-blue-800 text-white text-xs font-medium rounded-lg transition-all duration-200 min-h-[52px] transform active:scale-95"
-          >
-            <span className="text-lg mb-1">👁️</span>
-            <span>View</span>
-          </button>
-          
-          <button
-            onClick={handleEditClick}
-            className="flex flex-col items-center justify-center py-3 px-1 bg-green-600 hover:bg-green-700 active:bg-green-800 text-white text-xs font-medium rounded-lg transition-all duration-200 min-h-[52px] transform active:scale-95"
-          >
-            <span className="text-lg mb-1">✏️</span>
-            <span>Edit</span>
-          </button>
-          
-          <button
-            onClick={() => onProcessPayment && onProcessPayment(student)}
-            className="flex flex-col items-center justify-center py-3 px-1 bg-yellow-600 hover:bg-yellow-700 active:bg-yellow-800 text-white text-xs font-medium rounded-lg transition-all duration-200 min-h-[52px] transform active:scale-95"
-          >
-            <span className="text-lg mb-1">💳</span>
-            <span>Pay</span>
-          </button>
-
-          <button
-            onClick={handleTrainingClick}
-            className="flex flex-col items-center justify-center py-3 px-1 bg-orange-600 hover:bg-orange-700 active:bg-orange-800 text-white text-xs font-medium rounded-lg transition-all duration-200 min-h-[52px] transform active:scale-95"
-          >
-            <span className="text-lg mb-1">🥋</span>
-            <span>Training</span>
-          </button>
-        </div>
-        
-        {canSendSMS && (
-          <button
-            onClick={() => onSendReminder && onSendReminder(student)}
-            disabled={smsLoading}
-            className="w-full flex items-center justify-center py-3 px-4 bg-purple-600 hover:bg-purple-700 active:bg-purple-800 text-white text-sm font-medium rounded-lg transition-all duration-200 disabled:opacity-50 disabled:cursor-not-allowed min-h-[52px] transform active:scale-95"
-          >
-            <span className="mr-2 text-lg">{smsLoading ? "⏳" : "📱"}</span>
-            <span>{smsLoading ? "Sending SMS..." : "Send SMS Reminder"}</span>
-          </button>
-        )}
-      </div>
-    </div>
-  );
-};
-
-// Lines 130-205: Desktop Table Row Component - Refactored with extracted utilities
-const StudentTableRow = ({ 
-  student, 
-  onProcessPayment, 
-  onViewStudent, 
-  onEditStudent, 
-  onSendReminder, 
-  onLogTraining,
-  canSendReminder, 
-  smsLoading,
-  students = [],
-  getStudentStatus,
-  getDaysRemaining
-}) => {
-  if (!student || typeof student !== 'object') {
-    return null;
-  }
-
-  // Lines 145-150: Use utility function for data normalization
-  const studentData = normalizeStudentData(student);
-  const { name, email, phone, monthlyRate } = studentData;
-
-  // Lines 152-157: Status calculations
-  const status = getStudentStatus ? getStudentStatus(student) : 'inactive';
-  const daysText = getDaysRemaining ? getDaysRemaining(student) : 'Unknown';
-  const canSendSMS = canSendReminder ? canSendReminder(student) : false;
-
-  // Lines 159-167: Safe event handlers using utility
-  const handleEditClick = createSafeHandler(
-    () => onEditStudent(student), 
-    'Desktop Edit - onEditStudent'
-  );
-
-  const handleTrainingClick = createSafeHandler(
-    () => onLogTraining(student), 
-    'Desktop Training - onLogTraining'
-  );
-
-  return (
-    <tr className="hover:bg-gray-750 transition-colors duration-200">
-      <td className="px-6 py-4 whitespace-nowrap">
-        <div className="flex items-center">
-          <div className="flex-shrink-0 h-12 w-12">
-            <div className="h-12 w-12 rounded-full bg-gradient-to-r from-blue-400 to-purple-500 flex items-center justify-center text-white font-bold text-lg">
-              {name.charAt(0).toUpperCase()}
-            </div>
-          </div>
-          <div className="ml-4">
-            <div className="text-sm font-semibold text-white">
-              {name}
-            </div>
-            <div className="text-sm text-gray-400">
-              {email}
-            </div>
-            <div className="text-xs text-gray-500">
-              📞 {phone}
-            </div>
-          </div>
-        </div>
-      </td>
-      
-      <td className="px-6 py-4 whitespace-nowrap">
-        <StudentStatusDisplay 
-          status={status} 
-          student={student} 
-          showLegacy={true}
-        />
-      </td>
-      
-      <td className="px-6 py-4 whitespace-nowrap">
-        <div className="text-sm font-medium text-white">MONTHLY</div>
-        <div className="text-lg font-bold text-white">
-          ₱{monthlyRate.toLocaleString()}/mo
-        </div>
-      </td>
-      
-      <td className="px-6 py-4 whitespace-nowrap">
-        <span className="text-sm font-medium text-gray-300">
-          {daysText}
-        </span>
-      </td>
-      
-      <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
-        <div className="flex items-center justify-end space-x-2">
-          <button
-            onClick={() => onViewStudent && onViewStudent(student.id)}
-            className="text-blue-400 hover:text-blue-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200 min-h-[44px] min-w-[44px] flex items-center justify-center"
-            title="View Student Details"
-          >
-            <span className="text-lg">👁️</span>
-          </button>
-          
-          <button
-            onClick={handleEditClick}
-            className="text-green-400 hover:text-green-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200 min-h-[44px] min-w-[44px] flex items-center justify-center"
-            title="Edit Student Information"
-          >
-            <span className="text-lg">✏️</span>
-          </button>
-          
-          <button
-            onClick={() => onProcessPayment && onProcessPayment(student)}
-            className="text-yellow-400 hover:text-yellow-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200 min-h-[44px] min-w-[44px] flex items-center justify-center"
-            title="Process Payment"
-          >
-            <span className="text-lg">💳</span>
-          </button>
-
-          <button
-            onClick={handleTrainingClick}
-            className="text-orange-400 hover:text-orange-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200 min-h-[44px] min-w-[44px] flex items-center justify-center"
-            title="Log Training Session"
-          >
-            <span className="text-lg">🥋</span>
-          </button>
-          
-          {canSendSMS && (
-            <button
-              onClick={() => onSendReminder && onSendReminder(student)}
-              disabled={smsLoading}
-              className="text-purple-400 hover:text-purple-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200 disabled:opacity-50 disabled:cursor-not-allowed min-h-[44px] min-w-[44px] flex items-center justify-center"
-              title={`Send SMS Reminder to ${name}`}
-            >
-              <span className="text-lg">{smsLoading ? "⏳" : "📱"}</span>
-            </button>
-          )}
-        </div>
-      </td>
-    </tr>
-  );
-};
-
-// Lines 250-380: Main Component - Refactored with extracted utilities
+/**
+ * StudentManagementSection Component - Phase 2 Refactored
+ * Main container for student management with extracted sub-components
+ * Lines 15-20: Reduced to coordination and layout only
+ */
 const StudentManagementSection = ({
   filteredStudents,
   students,
@@ -287,18 +34,17 @@ const StudentManagementSection = ({
   getDaysRemaining,
   smsLoading = false
 }) => {
-  // Lines 275-280: Use utility functions
+  // Lines 35-40: Simplified data preparation
   const safeFilteredStudents = ensureArray(filteredStudents);
   const safeStudents = ensureArray(students);
   const safeTabCounts = tabCounts || {};
 
-  // Lines 282-287: Generate tabs with counts using configuration
   const tabs = tabsConfig.map(tab => ({
     ...tab,
     count: safeTabCounts[tab.id] || 0
   }));
 
-  // Lines 289-310: Safe event handlers (preserved functionality)
+  // Lines 45-65: Event handlers extracted to parent coordination
   const handleSearchChange = (e) => {
     const value = e.target.value;
     if (setSearchQuery && typeof setSearchQuery === 'function') {
@@ -318,8 +64,15 @@ const StudentManagementSection = ({
     }
   };
 
+  const handleTabChange = (tabId) => {
+    if (setCurrentTab && typeof setCurrentTab === 'function') {
+      setCurrentTab(tabId);
+    }
+  };
+
   return (
     <div className="bg-gray-800 shadow-xl rounded-xl overflow-hidden border border-gray-700">
+      {/* Header Section */}
       <div className="bg-gradient-to-r from-gray-700 to-gray-800 px-4 sm:px-6 py-4 border-b border-gray-700">
         <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-4">
           <div className="flex-1">
@@ -338,48 +91,23 @@ const StudentManagementSection = ({
         </div>
       </div>
 
+      {/* Search and Filter Controls */}
       <div className="px-4 sm:px-6 py-4 bg-gray-750 border-b border-gray-700 space-y-4">
-        <div className="relative">
-          <input
-            type="text"
-            placeholder="Search students by name, email, or phone..."
-            value={searchQuery || ''}
-            onChange={handleSearchChange}
-            className="block w-full pl-12 pr-12 py-3 border border-gray-600 rounded-lg bg-gray-700 text-white placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent text-base min-h-[52px] transition-all duration-200"
-          />
-          <div className="absolute inset-y-0 left-0 pl-4 flex items-center pointer-events-none">
-            <span className="text-gray-400 text-lg">🔍</span>
-          </div>
-          {isSearchActive && (
-            <button
-              onClick={handleClearSearch}
-              className="absolute inset-y-0 right-0 pr-4 flex items-center min-h-[52px] min-w-[52px] justify-center transition-colors duration-200"
-            >
-              <span className="text-gray-400 hover:text-white text-lg">❌</span>
-            </button>
-          )}
-        </div>
+        <SearchControls
+          searchQuery={searchQuery}
+          isSearchActive={isSearchActive}
+          onSearchChange={handleSearchChange}
+          onClearSearch={handleClearSearch}
+        />
 
-        <div className="flex gap-2 overflow-x-auto scrollbar-hide pb-2 px-1">
-          {tabs.map((tab) => (
-            <button
-              key={tab.id}
-              onClick={() => setCurrentTab && setCurrentTab(tab.id)}
-              className={`flex-shrink-0 px-4 py-2 rounded-lg text-sm font-medium transition-all duration-200 whitespace-nowrap min-h-[48px] transform active:scale-95 ${
-                currentTab === tab.id
-                  ? "bg-blue-600 text-white shadow-lg"
-                  : "bg-gray-700 text-gray-300 hover:text-white hover:bg-gray-600"
-              }`}
-            >
-              <div className="text-center">
-                <div className="font-medium">{tab.label}</div>
-                <div className="text-xs opacity-75">({tab.count})</div>
-              </div>
-            </button>
-          ))}
-        </div>
+        <TabNavigation
+          tabs={tabs}
+          currentTab={currentTab}
+          onTabChange={handleTabChange}
+        />
       </div>
 
+      {/* Status Bar */}
       <div className="px-4 sm:px-6 py-3 bg-gray-750 border-b border-gray-700">
         <div className="flex justify-between items-center text-sm">
           <span className="text-gray-300 font-medium">
@@ -394,6 +122,7 @@ const StudentManagementSection = ({
         </div>
       </div>
 
+      {/* Student Display - Mobile/Desktop */}
       {safeFilteredStudents.length > 0 ? (
         <>
           {/* Mobile Cards */}
@@ -403,7 +132,6 @@ const StudentManagementSection = ({
                 <StudentCard
                   key={student.id}
                   student={student}
-                  students={safeStudents}
                   onProcessPayment={onProcessPayment}
                   onViewStudent={onViewStudent}
                   onEditStudent={onEditStudent}
@@ -445,7 +173,6 @@ const StudentManagementSection = ({
                   <StudentTableRow
                     key={student.id}
                     student={student}
-                    students={safeStudents}
                     onProcessPayment={onProcessPayment}
                     onViewStudent={onViewStudent}
                     onEditStudent={onEditStudent}
