+++ client/src/components/student/StudentTableRow.jsx
@@ -1,352 +1,133 @@
 // File: client/src/components/student/StudentTableRow.jsx
-// Lines 1-20: FIXED STUDENT TABLE ROW - NaN Prevention and Safe Date Handling
+// Lines 1-25: Aligned with Phase 1 utilities
 import React from 'react';
+import { normalizeStudentData, createSafeHandler } from '../../utils/studentTableUtils';
+import { getStatusConfig } from '../../utils/studentStatusConfig';
+import StudentStatusDisplay from './StudentStatusDisplay';
 
 /**
- * StudentTableRow Component - PRODUCTION READY
- * Renders individual student table row with safe date calculations
- * 
- * CRITICAL FIXES APPLIED:
- * - Eliminated NaN days calculation errors
- * - Safe string operations with comprehensive validation
- * - Enhanced error handling for date parsing
- * - Proper membership selection logic
- * - No console logging for production
+ * StudentTableRow Component - Phase 1 Aligned
+ * Renders individual student table row using extracted utilities
+ * REMOVED: Inline utilities, embedded status logic, duplicate date calculations
  */
-
-// Lines 25-80: SAFE DATE CALCULATION UTILITIES - Inline to avoid import issues
-const safeDateParse = (dateInput) => {
-  if (dateInput === null || dateInput === undefined) {
-    return null;
-  }
-  
-  try {
-    if (dateInput instanceof Date) {
-      return isNaN(dateInput.getTime()) ? null : dateInput;
-    }
-    
-    const dateStr = String(dateInput).trim();
-    if (!dateStr || dateStr === 'null' || dateStr === 'undefined') {
-      return null;
-    }
-    
-    const parsedDate = new Date(dateStr);
-    return isNaN(parsedDate.getTime()) ? null : parsedDate;
-  } catch (error) {
-    return null;
-  }
-};
-
-const calculateSafeDaysDifference = (endDateInput, startDateInput = null) => {
-  try {
-    const endDate = safeDateParse(endDateInput);
-    if (!endDate) return null;
-    
-    const startDate = startDateInput ? safeDateParse(startDateInput) : new Date();
-    if (!startDate) return null;
-    
-    const endDateOnly = new Date(endDate.getFullYear(), endDate.getMonth(), endDate.getDate());
-    const startDateOnly = new Date(startDate.getFullYear(), startDate.getMonth(), startDate.getDate());
-    
-    const timeDiff = endDateOnly.getTime() - startDateOnly.getTime();
-    if (isNaN(timeDiff)) return null;
-    
-    const daysDiff = Math.ceil(timeDiff / (1000 * 60 * 60 * 24));
-    return isNaN(daysDiff) ? null : daysDiff;
-  } catch (error) {
-    return null;
-  }
-};
-
-const formatSafeDueDate = (dateInput) => {
-  if (!dateInput) {
-    return { text: "N/A", color: "text-gray-400" };
-  }
-  
-  try {
-    const daysDiff = calculateSafeDaysDifference(dateInput);
-    
-    if (daysDiff === null || isNaN(daysDiff)) {
-      return { text: "Invalid Date", color: "text-red-400" };
-    }
-    
-    if (daysDiff > 7) {
-      return { 
-        text: `${daysDiff} days remaining`, 
-        color: "text-green-400" 
-      };
-    } else if (daysDiff > 0) {
-      return { 
-        text: `${daysDiff} day${daysDiff === 1 ? '' : 's'} remaining`, 
-        color: "text-yellow-400" 
-      };
-    } else if (daysDiff === 0) {
-      return { 
-        text: "Due today", 
-        color: "text-orange-400 font-medium" 
-      };
-    } else {
-      const overdueDays = Math.abs(daysDiff);
-      return { 
-        text: `${overdueDays} day${overdueDays === 1 ? '' : 's'} overdue`, 
-        color: "text-red-400 font-medium" 
-      };
-    }
-  } catch (error) {
-    return { text: "Error", color: "text-red-400" };
-  }
-};
-
-// Lines 85-130: SAFE MEMBERSHIP HELPER FUNCTIONS
-const getSafeLatestMembership = (student) => {
-  if (!student || typeof student !== 'object') {
-    return null;
-  }
-
-  if (!student.memberships || !Array.isArray(student.memberships) || student.memberships.length === 0) {
-    return null;
-  }
-
-  try {
-    const sortedMemberships = [...student.memberships].sort((a, b) => {
-      const getDateValue = (membership) => {
-        const dateStr = membership.createdAt || membership.endDate || membership.startDate;
-        const date = safeDateParse(dateStr);
-        return date ? date.getTime() : 0;
-      };
-      
-      return getDateValue(b) - getDateValue(a); // DESC order - newest first
-    });
-
-    return sortedMemberships[0] || null;
-  } catch (error) {
-    return null;
-  }
-};
-
-const getSafeStudentStatus = (student) => {
-  try {
-    if (!student || typeof student !== 'object') {
-      return 'inactive';
-    }
-
-    const latestMembership = getSafeLatestMembership(student);
-    if (!latestMembership || !latestMembership.endDate) {
-      return 'inactive';
-    }
-
-    const daysDiff = calculateSafeDaysDifference(latestMembership.endDate);
-    
-    if (daysDiff === null || isNaN(daysDiff)) {
-      return 'inactive';
-    }
-
-    if (daysDiff < 0) return 'overdue';
-    if (daysDiff <= 7) return 'expiring';
-    return 'active';
-  } catch (error) {
-    return 'inactive';
-  }
-};
-
-// Lines 135-200: MAIN COMPONENT WITH SAFE RENDERING
 const StudentTableRow = ({ 
   student, 
   onProcessPayment, 
   onViewStudent, 
   onEditStudent, 
   onSendReminder, 
+  onLogTraining, // ADDED: Training integration
   canSendReminder, 
-  smsLoading = false,
+  smsLoading,
   getStudentStatus,
   getDaysRemaining
 }) => {
-  // Input validation
   if (!student || typeof student !== 'object') {
     return null;
   }
 
-  // Safe data extraction with fallbacks
-  const studentName = String(student.name || 'Unknown Student');
-  const studentEmail = String(student.email || 'No email');
-  const studentPhone = String(student.phone || student.phoneNumber || 'No phone');
-  const monthlyRate = parseFloat(student.monthlyRate || student.rate || 1400);
-  const isLegacy = student.isLegacyStudent || monthlyRate < 1400;
+  // Lines 30-35: Use Phase 1 utilities (not inline functions)
+  const studentData = normalizeStudentData(student);
+  const { name, email, phone, monthlyRate } = studentData;
 
-  // Use safe status calculation (prefer passed function, fallback to internal)
-  const status = getStudentStatus ? getStudentStatus(student) : getSafeStudentStatus(student);
-  
-  // Get latest membership safely
-  const latestMembership = getSafeLatestMembership(student);
-  
-  // Calculate due date info safely
-  const dueDateInfo = latestMembership?.endDate 
-    ? formatSafeDueDate(latestMembership.endDate)
-    : { text: "No membership", color: "text-gray-400" };
+  // Lines 37-42: Status calculations using passed functions
+  const status = getStudentStatus ? getStudentStatus(student) : 'inactive';
+  const daysText = getDaysRemaining ? getDaysRemaining(student) : 'Unknown';
+  const canSendSMS = canSendReminder ? canSendReminder(student) : false;
 
-  // SMS eligibility check
-  const canSendSMS = canSendReminder ? canSendReminder(student) : (
-    (status === 'expiring' || status === 'overdue') && 
-    Boolean(student.phone || student.phoneNumber)
+  // Lines 44-52: Safe event handlers using Phase 1 utilities
+  const handleEditClick = createSafeHandler(
+    () => onEditStudent(student), 
+    'Desktop Edit - onEditStudent'
   );
 
-  // Status badge configuration
-  const getStatusBadge = (status) => {
-    const statusConfig = {
-      active: { bg: "bg-green-600", text: "text-white", label: "✅ Active", icon: "✅" },
-      expiring: { bg: "bg-yellow-600", text: "text-white", label: "⚠️ Expiring", icon: "⚠️" },
-      overdue: { bg: "bg-red-600", text: "text-white", label: "🚨 Overdue", icon: "🚨" },
-      inactive: { bg: "bg-gray-600", text: "text-white", label: "⚫ Inactive", icon: "⚫" }
-    };
-
-    const config = statusConfig[status] || statusConfig.inactive;
-
-    return (
-      <div className="flex items-center space-x-2">
-        <span className={`inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium ${config.bg} ${config.text}`}>
-          <span className="mr-1">{config.icon}</span>
-          {config.label.replace(/[^a-zA-Z\s]/g, '')}
-        </span>
-        {isLegacy && (
-          <span className="inline-flex items-center px-2 py-0.5 rounded-full text-xs font-medium bg-purple-600 text-white">
-            ⭐ Legacy
-          </span>
-        )}
-      </div>
-    );
-  };
-
-  // Safe currency formatting
-  const formatCurrency = (amount) => {
-    if (!amount && amount !== 0) return "₱0";
-    try {
-      const numAmount = parseFloat(amount);
-      return isNaN(numAmount) ? "₱0" : `₱${numAmount.toLocaleString()}`;
-    } catch {
-      return "₱0";
-    }
-  };
-
-  // Safe date formatting
-  const formatDisplayDate = (dateStr) => {
-    if (!dateStr) return "N/A";
-    try {
-      const date = safeDateParse(dateStr);
-      return date ? date.toLocaleDateString() : "Invalid Date";
-    } catch {
-      return "Invalid Date";
-    }
-  };
+  const handleTrainingClick = createSafeHandler(
+    () => onLogTraining(student), 
+    'Desktop Training - onLogTraining'
+  );
 
   return (
-    <tr className="hover:bg-gray-750 transition-colors">
-      {/* Student Information Cell */}
+    <tr className="hover:bg-gray-750 transition-colors duration-200">
       <td className="px-6 py-4 whitespace-nowrap">
         <div className="flex items-center">
-          <div className="flex-shrink-0 h-10 w-10">
-            <div className="h-10 w-10 rounded-full bg-gradient-to-r from-blue-400 to-purple-500 flex items-center justify-center text-white font-semibold">
-              {studentName.charAt(0).toUpperCase()}
+          <div className="flex-shrink-0 h-12 w-12">
+            <div className="h-12 w-12 rounded-full bg-gradient-to-r from-blue-400 to-purple-500 flex items-center justify-center text-white font-bold text-lg">
+              {name.charAt(0).toUpperCase()}
             </div>
           </div>
           <div className="ml-4">
-            <div className="text-sm font-medium text-white">
-              {studentName}
-            </div>
-            <div className="text-sm text-gray-400">
-              {studentEmail}
-            </div>
-            <div className="text-xs text-gray-500">
-              {studentPhone}
-            </div>
+            <div className="text-sm font-semibold text-white">{name}</div>
+            <div className="text-sm text-gray-400">{email}</div>
+            <div className="text-xs text-gray-500">📞 {phone}</div>
           </div>
         </div>
       </td>
       
-      {/* Status Cell */}
       <td className="px-6 py-4 whitespace-nowrap">
-        {getStatusBadge(status)}
+        <StudentStatusDisplay 
+          status={status} 
+          student={student} 
+          showLegacy={true}
+        />
       </td>
       
-      {/* Membership Cell */}
       <td className="px-6 py-4 whitespace-nowrap">
-        <div className="text-sm text-white">MONTHLY</div>
-        <div className="text-sm text-gray-400">
-          {formatCurrency(monthlyRate)}/mo
+        <div className="text-sm font-medium text-white">MONTHLY</div>
+        <div className="text-lg font-bold text-white">
+          ₱{monthlyRate.toLocaleString()}/mo
         </div>
-        {latestMembership?.startDate && (
-          <div className="text-xs text-gray-500">
-            Started: {formatDisplayDate(latestMembership.startDate)}
-          </div>
-        )}
       </td>
       
-      {/* Due Date Cell - FIXED: No more NaN errors */}
       <td className="px-6 py-4 whitespace-nowrap">
-        <span className={`text-sm font-medium ${dueDateInfo.color}`}>
-          {dueDateInfo.text}
+        <span className="text-sm font-medium text-gray-300">
+          {daysText}
         </span>
-        {latestMembership?.endDate && (
-          <div className="text-xs text-gray-500">
-            End: {formatDisplayDate(latestMembership.endDate)}
-          </div>
-        )}
       </td>
       
-      {/* Actions Cell */}
       <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
         <div className="flex items-center justify-end space-x-2">
-          {/* View Button */}
           <button
             onClick={() => onViewStudent && onViewStudent(student.id)}
-            className="text-blue-400 hover:text-blue-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200"
+            className="text-blue-400 hover:text-blue-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200 min-h-[44px] min-w-[44px] flex items-center justify-center"
             title="View Student Details"
           >
-            👁️
+            <span className="text-lg">👁️</span>
           </button>
           
-          {/* Edit Button */}
           <button
-            onClick={() => onEditStudent && onEditStudent(student)}
-            className="text-green-400 hover:text-green-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200"
+            onClick={handleEditClick}
+            className="text-green-400 hover:text-green-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200 min-h-[44px] min-w-[44px] flex items-center justify-center"
             title="Edit Student Information"
           >
-            ✏️
+            <span className="text-lg">✏️</span>
           </button>
           
-          {/* Payment Button */}
           <button
             onClick={() => onProcessPayment && onProcessPayment(student)}
-            className="text-yellow-400 hover:text-yellow-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200"
+            className="text-yellow-400 hover:text-yellow-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200 min-h-[44px] min-w-[44px] flex items-center justify-center"
             title="Process Payment"
           >
-            💳
+            <span className="text-lg">💳</span>
+          </button>
+
+          <button
+            onClick={handleTrainingClick}
+            className="text-orange-400 hover:text-orange-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200 min-h-[44px] min-w-[44px] flex items-center justify-center"
+            title="Log Training Session"
+          >
+            <span className="text-lg">🥋</span>
           </button>
           
-          {/* SMS Reminder Button (conditional) - FIXED: Proper eligibility check */}
           {canSendSMS && (
             <button
               onClick={() => onSendReminder && onSendReminder(student)}
               disabled={smsLoading}
-              className="text-purple-400 hover:text-purple-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200 disabled:opacity-50 disabled:cursor-not-allowed"
-              title={`Send SMS Reminder to ${studentName}`}
+              className="text-purple-400 hover:text-purple-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200 disabled:opacity-50 disabled:cursor-not-allowed min-h-[44px] min-w-[44px] flex items-center justify-center"
+              title={`Send SMS Reminder to ${name}`}
             >
-              {smsLoading ? "⏳" : "📱"}
+              <span className="text-lg">{smsLoading ? "⏳" : "📱"}</span>
             </button>
           )}
-          
-          {/* Delete/Remove Button */}
-          <button
-            onClick={() => {
-              if (window.confirm(`Are you sure you want to remove ${studentName}?`)) {
-                // Handle delete if needed
-                console.log('Delete student:', student.id);
-              }
-            }}
-            className="text-red-400 hover:text-red-300 p-2 rounded-lg hover:bg-gray-700 transition-all duration-200"
-            title="Remove Student"
-          >
-            🗑️
-          </button>
         </div>
       </td>
     </tr>
